requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
openai==1.6.1
pydantic==2.5.2
python-dotenv==1.0.0
typing-extensions==4.8.0
